requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
cloudinary==1.44.1
dj-database-url==3.0.1
Django==5.2.5
django-cloudinary-storage==0.3.0
django-parler==2.3
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.9.0
